provider/bluecat: migrate client code to separate pkg
Extract all Bluecat Gateway Client code to separate package
